Moved the index.js to lib/rtm.js
I moved and tidied up the index.js file a little bit.

The index.js file now imports the lib/rtm.js file which holds the contents of index.js
